Add support to specify timeout by ptest
The ptest-runner has a default timeout of 300 secs and can be override
usint -t option in the runner.

There is a need to specify timeout by ptest because not all ptests takes
the sametime also are affected by machine. So add support to read a file
inside ptest folder (timeout-ptest) and override global timeout with it.

For example in glib-2.0,

  /usr/lib/glib-2.0/ptest/run-ptest
  /usr/lib/glib-2.0/ptest/timeout-ptest

Signed-off-by: Aníbal Limón <anibal.limon@linaro.org>