clang: clean-ups to avoid -Weverything warnings.
The function cleanup_ptest_opts() is used at file scope so make it static to
avoid a warning about it not having a prototype.

Add padding variables to ensure that struct elements are word aligned
on 64 bit machines to avoid clang warnings such as:
   padding struct 'struct ptest_options' with 4 bytes to align 'exclude' [-Wpadded]

Make the 'timeout' variable types be 'unsigned int' since they are used with
alarm() and will otherwise generate a warning.

Explicitly avoid warnings such as:
   warning: unused parameter 'arg' [-Wunused-parameter]

Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>
Signed-off-by: Aníbal Limón <anibal.limon@linaro.org>